+++ image/etc_dojo.d/scripts/90-k8s.sh
@@ -1,0 +1,5 @@
+#!/bin/bash -e
+
+if [ -n "${AIT_GPD_K8S_USER}" ];  then
+  sudo -E -H -u dojo /usr/bin/setup-kube "${AIT_GPD_K8S_USER}"
+fi